requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
streamlit==1.35.0
python-dotenv==1.0.1

# ── Google Gemini ────────────────────────────────────────────────────────────
google-generativeai==0.7.2

# ── Deep Learning ────────────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0

# ── Transformers / CLIP / HuggingFace ────────────────────────────────────────
transformers==4.41.2
sentence-transformers==3.0.1
open-clip-torch==2.24.0
huggingface-hub==0.23.4

# ── Classical ML ─────────────────────────────────────────────────────────────
scikit-learn==1.5.0
xgboost==2.0.3

# ── Computer Vision ──────────────────────────────────────────────────────────
opencv-python-headless==4.9.0.80
Pillow==10.3.0
albumentations==1.4.8

# ── NLP ──────────────────────────────────────────────────────────────────────
nltk==3.8.1
keybert==0.8.4

# ── Data Processing ──────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# ── Visualization ────────────────────────────────────────────────────────────
matplotlib==3.9.0
plotly==5.22.0

# ── Recommendation System ────────────────────────────────────────────────────
faiss-cpu==1.8.0

# ── Color Science ────────────────────────────────────────────────────────────
colorthief==0.2.1

# ── Product Search ───────────────────────────────────────────────────────────
requests==2.31.0
serpapi==0.1.5
beautifulsoup4==4.12.3

# ── PDF Export ───────────────────────────────────────────────────────────────
fpdf2==2.7.9

# ── Utilities ────────────────────────────────────────────────────────────────
tqdm==4.66.4
joblib==1.4.2
cachetools==5.3.3
